+++ crates/abq_workers/src/lib.rs
@@ -1,3 +1,4 @@
+mod assigned_run;
 mod liveness;
 pub mod negotiate;
 pub mod results_handler;
@@ -8,3 +9,4 @@ pub mod workers;
 
 pub use abq_generic_test_runner::DEFAULT_PROTOCOL_VERSION_TIMEOUT;
 pub use abq_generic_test_runner::DEFAULT_RUNNER_TEST_TIMEOUT;
+pub use assigned_run::{AssignedRun, AssignedRunStatus, GetAssignedRun};
